ml-sentiment-app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
transformers==4.35.2
torch==2.1.1
pydantic==2.5.2
python-multipart==0.0.6
mlflow==2.8.1
docker==6.1.3
google-cloud-storage==2.13.0
python-dotenv==1.0.0 